build(nx): move dynamic config out of normal packages
