+++ packages/@clc/nx/package.json
@@ -1,0 +1,5 @@
+{
+  "name": "@clc/nx",
+  "main": "./src/index.ts",
+  "private": true
+}
